rewrite depends/imports etc
I don't think the importFrom roxygen2 things really work.  Yes, they
create items in the NAMESPACE, but then when I run ctdTag(), I get
errors.  This is not the case if I do it within a long RStudio session,
so I only just noticed this now, because I am now doing

    cd git
    make ctd
    Rscript 0.R

where 0.R is as below

    library(ctd)
    file <- "~/data/arctic/beaufort/2012/d201211_0047.cnv"
    ctdTag(file)