make database be more 'relational'
This change should make it easier to designate filenames as having been
processed, but not having any tagged features.  But it's more than that.
I am now using a 'relational' database schema.

Previously, `tags` had a column for filename and a column for tag label,
but the new code relies instead on use cross-references to other tables.
These cross-references are what makes the new scheme 'relational'.

To maintain compatibility with databases created with older versions of
the code, the tables are converted as required, if the version number in
the file differs from that in the code.

I have checked all of this fairly extensively, but there may still be
bugs.

Here is a checklist of changes (for my own organization, since I was
using `git commit --amend` for a series of perhaps a dozen steppingstone
commits).

* [x] make tables auto-update
* [x] make 'summary' use updated tables (requires merging)
* [x] show existing tags on plots
* [x] permit deleting existing tags
* [x] permit adding new tags
* [x] do not store tagLabel in `tags`